Deleting a peer permanently doesn't seem to be possible despite deleting the .json file and peer file under /var/lib/zerotier so the IP for the peer is set to 127.0.0.100 and the bash script ignores peers returned with that IP so it doesn't appear in the console.
A new menu item was added under "Peer Management" to:

'Undelete' a peer

a user of bash_cli_zt noted they 'deleted' a peer by accident
and it wouldn't show back up in the menu options for the
reason above.

The 'Undelete' option allows listing all deleted peers,
since those are always present, and allowing it to show
up in the peer list to reauthorize.

I've tried all the methods I could find online but
couldn't permanently delete the peers so this 'hack'
is the best solution for now.